% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_build.R
\name{all_languages}
\alias{all_languages}
\title{Return all survey languages as a character vector (primary language first)}
\usage{
all_languages(s)
}
\description{
Return all survey languages as a character vector (primary language first)
}
\keyword{internal}
