% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_build.R
\name{build_lang_rows}
\alias{build_lang_rows}
\title{Build SL-rows (language-specific survey settings) for one language}
\usage{
build_lang_rows(s, lang)
}
\description{
Build SL-rows (language-specific survey settings) for one language
}
\keyword{internal}
