% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_build.R
\name{build_quota_rows}
\alias{build_quota_rows}
\title{Build QTA and QTALS rows (appended at the end of the TSV)}
\usage{
build_quota_rows(quotas, langs, primary_lang, quota_id_map)
}
\description{
Build QTA and QTALS rows (appended at the end of the TSV)
}
\keyword{internal}
