% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_build.R
\name{build_structure_rows}
\alias{build_structure_rows}
\title{Build all G/Q/SQ/A/QTAM rows for one language block}
\usage{
build_structure_rows(
  structure,
  lang,
  id_map,
  primary_lang,
  quotas = NULL,
  quota_id_map = NULL
)
}
\description{
Build all G/Q/SQ/A/QTAM rows for one language block
}
\keyword{internal}
