% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_validate.R
\name{find_group_of_qst}
\alias{find_group_of_qst}
\title{Find the group name containing a given question code}
\usage{
find_group_of_qst(qst_code, structure)
}
\description{
Find the group name containing a given question code
}
\keyword{internal}
