% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_quota.R
\name{grow_quota}
\alias{grow_quota}
\title{Load and Prepare Survey Seed}
\usage{
grow_quota(tree, sprout = sprout)
}
\arguments{
\item{tree}{survey object}

\item{sprout}{list from imported YAML file created by \code{lime_sprout} or manually created}
}
\description{
Load and Prepare Survey Seed
}
