% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lime_sprout.R
\name{lime_sprout}
\alias{lime_sprout}
\title{Load and Prepare Survey Seed}
\usage{
lime_sprout(seed)
}
\arguments{
\item{seed}{\enumerate{
\item path to single file containing all (necessary) elements (settings, structure)
\item path to folder containing all the (necessary) files (named accordingly)
\item named list containing paths to files for (necessary) elements
\item named list containing paths to files and (pre-loaded or created) elements as list
\item (pre-loaded or created) list with all (necessary) elements
}}
}
\description{
Load and Prepare Survey Seed
}
