% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_build.R
\name{ls_row}
\alias{ls_row}
\title{Create a named list for one TSV row (all values coerced to character)}
\usage{
ls_row(...)
}
\description{
Create a named list for one TSV row (all values coerced to character)
}
\keyword{internal}
