% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_build.R
\name{resolve_settings}
\alias{resolve_settings}
\title{Merge user settings over package defaults}
\usage{
resolve_settings(settings)
}
\description{
Merge user settings over package defaults
}
\keyword{internal}
