% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_validate.R
\name{validate_quota}
\alias{validate_quota}
\title{Validate Quota definitions in the LimeSeed sprout}
\usage{
validate_quota(quotas, structure, add)
}
\arguments{
\item{quotas}{The \verb{$quota} list from the parsed YAML.}

\item{structure}{The \verb{$structure} list from the parsed YAML.}

\item{add}{The internal \code{add(severity, location, msg)} closure.}
}
\description{
Validate Quota definitions in the LimeSeed sprout
}
\keyword{internal}
